# src/documentation/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "HLTV API Rest",
    "description": "<b>Unofficial</b> API Rest for <a href=\"https://www.hltv.org/\">HLTV</a>."
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Local Server API"
    },
    {
      "url": "https://hltv-api.onrender.com/api/v1",
      "description": "Production Server API"
    }
  ],
...
  "paths": {
    "/search": {
      "get": {
        "tags": ["General"],
        "summary": "General search in HLTV",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/GeneralSearch"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/ranking/{year}/{month}/{day}/{country}": {
      "get": {
        "tags": ["General"],
        "summary": "Ranking of bests",
        "parameters": [
          {
            "name": "year",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "month",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "day",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "country",
            "in": "path",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Ranking"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Ranking": {
        "type": "object",
        "properties": {
          "teams": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "position": {
                  "type": "object",
                  "properties": {
                    "actual": {
                      "type": "integer"
                    },
                    "change": {
                      "type": "integer"
                    }
                  }
                },
                "name": {
                  "type": "string"
                },
                "link": {
                  "type": "string"
                },
                "points": {
                  "type": "integer"
                },
                "players": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "full_name": {
                        "type": "string"
                      },
                      "nick": {
                        "type": "string"
                      },
                      "image": {
                        "type": "string"
                      },
                      "nationality": {
                        "type": "object",
                        "properties": {
                          "name": {
                            "type": "string"
                          },
                          "flag": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "navigation": {
            "type": "object",
            "properties": {
              "prev": {
                "type": "string",
                "nullable": true
              },
              "next": {
                "type": "string",
                "nullable": true
              }
            }
          }
        }
      },
      "Search": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "logo": {
            "type": "string"
          }
        }
      },
      "GeneralSearch": {
        "type": "object",
        "properties": {
          "search": {
            "type": "object",
            "properties": {
              "players": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Search"
                }
              },
              "teams": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Search"
                }
              },
              "articles": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Search"
                }
              },
              "events": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Search"
                }
              }
            }
          }
        }
      }
    }
  }
}